status_reacts: moved to util folder
